fix(async_hooks): integrate latest main
