+++ crates/perry-stdlib/Cargo.toml
@@ -228,6 +228,7 @@ external-tls-server = [
     "dep:rustls-native-certs",
     "dep:rustls-pemfile",
     "dep:base64",
+    "dep:x509-cert",
 ]
 
 # Databases
