+++ crates/perry/src/commands/compile/optimized_libs/driver.rs
@@ -460,6 +460,13 @@ pub(crate) fn build_optimized_libs(
             // them would drop unresolved externs at link time.
             if matches!(module_normalized, "http" | "https" | "http2") {
                 features.insert("external-http-server-pump");
+                // perry-ext-http depends on perry-ext-net, whose TLS client
+                // object calls back into perry-stdlib's main-thread
+                // preflight hook. Retain that provider even for plain HTTP:
+                // the static archive can pull the shared socket object before
+                // a TLS-specific API is used, and otherwise leaves
+                // `js_tls_client_preflight` undefined at link time.
+                features.insert("external-tls-server");
             }
             // Issue #769 — when `node:http` / `node:https` routes to
             // perry-ext-http, also activate the client-side pump so the
